+++ docs/source/change_log.rst
@@ -7,6 +7,11 @@ Phisher.
 Version 1.x.x
 -------------
 
+Version 1.4.0
+^^^^^^^^^^^^^
+
+*In Progress*
+
 Version 1.3.0
 ^^^^^^^^^^^^^